requirements.txt
# Core
openai>=1.0.0          # EigenAI uses OpenAI-compatible API
pydantic>=2.0
pyyaml>=6.0
python-dotenv>=1.0
httpx>=0.27.0

# Ingestion
feedparser>=6.0        # RSS feed parsing
newspaper4k>=0.9       # Article extraction

# Blockchain / Attestation
web3>=7.0              # Ethereum interaction
eth-account>=0.13      # Wallet management

# Web UI
fastapi>=0.115
uvicorn>=0.32

# Observability
structlog>=24.1
